requirements.txt
requests>=2.3.0
easygui
PySimpleGUI>=4.47.0
numpy
reportlab
anytree
plotly
pandas
google-cloud-documentai==2.14.0
